Move middleware to support directory/namespace
Update tests accordingly.
